+++ pytest_reportportal/config.py
@@ -32,6 +32,7 @@ except ImportError:
 class AgentConfig:
     """Storage for the RP agent initialization attributes."""
 
+    rp_enabled: bool
     rp_client_type: Optional[ClientType]
     rp_rerun: Optional[bool]
     pconfig: Config
@@ -53,7 +54,7 @@ class AgentConfig:
     rp_tests_attributes: Optional[List[str]]
     rp_launch_description: str
     rp_log_batch_size: int
-    rp_log_batch_payload_size: int
+    rp_log_batch_payload_limit: int
     rp_log_level: Optional[int]
     rp_log_format: Optional[str]
     rp_mode: str
@@ -61,8 +62,18 @@ class AgentConfig:
     rp_project: str
     rp_rerun_of: Optional[str]
     rp_api_retries: int
-    rp_skip_connection_test: bool
-    rp_api_key: str
+
+    # API key auth parameter
+    rp_api_key: Optional[str]
+
+    # OAuth 2.0 parameters
+    rp_oauth_uri: Optional[str]
+    rp_oauth_username: Optional[str]
+    rp_oauth_password: Optional[str]
+    rp_oauth_client_id: Optional[str]
+    rp_oauth_client_secret: Optional[str]
+    rp_oauth_scope: Optional[str]
+
     rp_verify_ssl: Union[bool, str]
     rp_launch_timeout: int
     rp_launch_uuid_print: bool
@@ -72,8 +83,9 @@ class AgentConfig:
 
     def __init__(self, pytest_config: Config) -> None:
         """Initialize required attributes."""
+        self.rp_enabled = to_bool(getattr(pytest_config.option, "rp_enabled", True))
         self.rp_rerun = pytest_config.option.rp_rerun or pytest_config.getini("rp_rerun")
-        self.rp_endpoint = self.find_option(pytest_config, "rp_endpoint")
+        self.rp_endpoint = getenv("RP_ENDPOINT") or self.find_option(pytest_config, "rp_endpoint")
         self.rp_hierarchy_code = to_bool(self.find_option(pytest_config, "rp_hierarchy_code"))
         self.rp_dir_level = int(self.find_option(pytest_config, "rp_hierarchy_dirs_level"))
         self.rp_hierarchy_dirs = to_bool(self.find_option(pytest_config, "rp_hierarchy_dirs"))
@@ -100,11 +112,11 @@ class AgentConfig:
         self.rp_tests_attributes = self.find_option(pytest_config, "rp_tests_attributes")
         self.rp_launch_description = self.find_option(pytest_config, "rp_launch_description")
         self.rp_log_batch_size = int(self.find_option(pytest_config, "rp_log_batch_size"))
-        batch_payload_size = self.find_option(pytest_config, "rp_log_batch_payload_size")
-        if batch_payload_size:
-            self.rp_log_batch_payload_size = int(batch_payload_size)
+        batch_payload_size_limit = self.find_option(pytest_config, "rp_log_batch_payload_limit")
+        if batch_payload_size_limit:
+            self.rp_log_batch_payload_limit = int(batch_payload_size_limit)
         else:
-            self.rp_log_batch_payload_size = MAX_LOG_BATCH_PAYLOAD_SIZE
+            self.rp_log_batch_payload_limit = MAX_LOG_BATCH_PAYLOAD_SIZE
         self.rp_log_level = get_actual_log_level(pytest_config, "rp_log_level")
         self.rp_log_format = self.find_option(pytest_config, "rp_log_format")
         self.rp_thread_logging = to_bool(self.find_option(pytest_config, "rp_thread_logging") or False)
@@ -112,7 +124,6 @@ class AgentConfig:
         self.rp_parent_item_id = self.find_option(pytest_config, "rp_parent_item_id")
         self.rp_project = self.find_option(pytest_config, "rp_project")
         self.rp_rerun_of = self.find_option(pytest_config, "rp_rerun_of")
-        self.rp_skip_connection_test = to_bool(self.find_option(pytest_config, "rp_skip_connection_test"))
 
         rp_api_retries_str = self.find_option(pytest_config, "rp_api_retries")
         rp_api_retries = rp_api_retries_str and int(rp_api_retries_str)
@@ -134,27 +145,16 @@ class AgentConfig:
             else:
                 self.rp_api_retries = 0
 
+        # API key auth parameter
         self.rp_api_key = getenv("RP_API_KEY") or self.find_option(pytest_config, "rp_api_key")
-        if not self.rp_api_key:
-            self.rp_api_key = getenv("RP_UUID") or self.find_option(pytest_config, "rp_uuid")
-            if self.rp_api_key:
-                warnings.warn(
-                    "Parameter `rp_uuid` is deprecated since 5.1.9 "
-                    "and will be subject for removing in the next "
-                    "major version. Use `rp_api_key` argument "
-                    "instead.",
-                    DeprecationWarning,
-                    2,
-                )
-            else:
-                warnings.warn(
-                    "Argument `rp_api_key` is `None` or empty string, "
-                    "that is not supposed to happen because Report "
-                    "Portal is usually requires an authorization key. "
-                    "Please check your configuration.",
-                    RuntimeWarning,
-                    2,
-                )
+
+        # OAuth 2.0 parameters
+        self.rp_oauth_uri = self.find_option(pytest_config, "rp_oauth_uri")
+        self.rp_oauth_username = self.find_option(pytest_config, "rp_oauth_username")
+        self.rp_oauth_password = self.find_option(pytest_config, "rp_oauth_password")
+        self.rp_oauth_client_id = self.find_option(pytest_config, "rp_oauth_client_id")
+        self.rp_oauth_client_secret = self.find_option(pytest_config, "rp_oauth_client_secret")
+        self.rp_oauth_scope = self.find_option(pytest_config, "rp_oauth_scope")
 
         rp_verify_ssl = self.find_option(pytest_config, "rp_verify_ssl", True)
         try: