+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- OAuth 2.0 Password Grant authentication, by @HardNorth
+### Changed
+- Client version updated to [5.6.7](https://github.com/reportportal/client-Python/releases/tag/5.6.7), by @HardNorth
+### Fixed
+- Some configuration parameter names, which are different in the client, by @HardNorth
+### Removed
+- `rp_uuid` param support, as it was deprecated pretty while ago, by @HardNorth
+
+## [5.5.2]
 ### Fixed
 - Issue [#397](https://github.com/reportportal/agent-python-pytest/issues/397) pytest launch stuck with xdist, by @HardNorth